todo-web-app/services/recurring-task/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dateutil==2.9.0
dapr==1.12.0
dapr-ext-fastapi==1.12.0
pydantic==2.12.5
httpx==0.28.1